update FilterRepository - перенес всю логику сохранения и получения данных в клас шаред преф
