+++ app/src/main/java/ru/practicum/android/diploma/ui/filter/workplace/country/CountryFragment.kt
@@ -59,5 +59,4 @@ class CountryFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
-
 }
